+++ test/patch.js
@@ -1,8 +1,79 @@
 
 var connect = require('../');
+var utils = require('../lib/utils');
 
 describe('patch', function(){
   describe('res', function(){
+    describe('cookie', function(){
+      it('should set a cookie', function(done){
+        var app = connect();
+
+        app.use(function(req, res){
+          res.cookie('name', 'tobi');
+          res.end();
+        });
+
+        app.request()
+        .get('/')
+        .end(function(res){
+          res.headers['set-cookie'].should.eql(['name=tobi; Path=/']);
+          done();
+        })
+      })
+
+      it('should accept options', function(done){
+        var app = connect();
+
+        app.use(function(req, res){
+          res.cookie('name', 'tobi', { httpOnly: true, secure: true });
+          res.end();
+        });
+
+        app.request()
+        .get('/')
+        .end(function(res){
+          res.headers['set-cookie'].should.eql(['name=tobi; Path=/; HttpOnly; Secure']);
+          done();
+        })
+      })
+
+      it('should accept maxAge', function(done){
+        var app = connect();
+
+        app.use(function(req, res){
+          res.cookie('name', 'tobi', { maxAge: 1000 });
+          res.end();
+        });
+
+        app.request()
+        .get('/')
+        .end(function(res){
+          res.headers['set-cookie'][0].should.not.include('Thu, 01 Jan 1970 00:00:01 GMT');
+          res.headers['set-cookie'][0].should.include('Max-Age=1');
+          done();
+        })
+      })
+
+      it('should not mutate the options object', function(done){
+        var app = connect();
+
+        var options = { maxAge: 1000 };
+        var optionsCopy = utils.merge({}, options);
+
+        app.use(function(req, res){
+          res.cookie('name', 'tobi', options)
+          res.end();
+        });
+
+        app.request()
+        .get('/')
+        .end(function(res){
+          options.should.eql(optionsCopy);
+          done();
+        })
+      })
+    })
+
     describe('headerSent', function(){
       it('should match res._header Boolean status', function(done){
         var app = connect();
@@ -165,7 +236,9 @@ describe('patch', function(){
         });
       })
     })
+  })
 
+  describe('res.setHeader', function(){
     describe('should handle set-cookie setting multiple cookies as array', function(){
       it('should be emitted', function(done){
         var app = connect();