+++ History.md
@@ -1,6 +1,7 @@
 HEAD
 ====
 
+  * Add simple `res.cookie` support
   * Patch `res.headerSent` to return Boolean
   * Patch `res.headersSent` for node.js 0.8
   * dep: connect-timeout@1.0.1